Add proper tests
